feat: export CLI utilities
