+++ package.json
@@ -10,6 +10,10 @@
       "types": "./dist/qdl.d.ts",
       "import": "./dist/qdl.js"
     },
+    "./cli": {
+      "types": "./dist/cli.d.ts",
+      "import": "./dist/cli.js"
+    },
     "./usblib": {
       "types": "./dist/usblib.d.ts",
       "import": "./dist/usblib.js"
